feat: ad h2 db and other things according to the task list

+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/AlreadyExistsException.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.exceptions;
 
 public class AlreadyExistsException extends RuntimeException {
 
-    public AlreadyExistsException(String objectName, Long id) {
+    public AlreadyExistsException(String objectName, Integer id) {
         super(String.format("%s с таким id(%d) уже существет!", objectName, id));
     }
 }
